Fix varargs in lev_comp so this works on 64-bit systems too.
git-svn-id: svn://sporkhack.com/nh@543 0b9a868d-f130-0410-bc2a-adc7ae8015a6

+++ util/lev_main.c
@@ -359,7 +359,7 @@ add_opvars(sp_lev *sp, const char *fmt, ...)
 	    }
 	case 'o':
 	    {
-		long i = va_arg(argp, long);
+		long i = va_arg(argp, int);
 		if (i < 0 || i >= MAX_SP_OPCODES)
 		    fprintf(stderr, "add_opvars: unknown opcode '%i'.", i);
 		add_opcode(sp, i, NULL);
